move DocumentJoiner to new joiners package
