+++ e2e/pipelines/test_eval_hybrid_doc_search_pipeline.py
@@ -2,7 +2,7 @@ from haystack import Document, Pipeline
 from haystack.components.embedders import SentenceTransformersTextEmbedder
 from haystack.components.rankers import TransformersSimilarityRanker
 from haystack.components.retrievers import InMemoryBM25Retriever, InMemoryEmbeddingRetriever
-from haystack.components.routers.document_joiner import DocumentJoiner
+from haystack.components.joiners.document_joiner import DocumentJoiner
 from haystack.document_stores import InMemoryDocumentStore
 from haystack.evaluation.eval import eval
