+++ haystack/components/joiners/__init__.py
@@ -1,0 +1,3 @@
+from haystack.components.joiners.document_joiner import DocumentJoiner
+
+__all__ = ["DocumentJoiner"]
