+++ haystack/components/joiners/document_joiner.py
