+++ haystack/components/routers/__init__.py
@@ -1,7 +1,6 @@
-from haystack.components.routers.document_joiner import DocumentJoiner
 from haystack.components.routers.file_type_router import FileTypeRouter
 from haystack.components.routers.metadata_router import MetadataRouter
 from haystack.components.routers.conditional_router import ConditionalRouter
 from haystack.components.routers.text_language_router import TextLanguageRouter
 
-__all__ = ["DocumentJoiner", "FileTypeRouter", "MetadataRouter", "TextLanguageRouter", "ConditionalRouter"]
+__all__ = ["FileTypeRouter", "MetadataRouter", "TextLanguageRouter", "ConditionalRouter"]
